backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
python-multipart>=0.0.6
python-pptx>=0.6.21
typing-extensions>=4.8.0
